upexam/requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.0.6
django-autoslug==1.9.9
django-ckeditor==6.4.2
django-cors-headers==3.13.0
django-environ==0.9.0
django-grappelli==3.0.8
django-import-export==3.3.1
django-js-asset==2.0.0
django-redis==5.2.0
django-storages==1.12.3
djangorestframework==3.13.1
et-xmlfile==1.1.0
idna==3.6
itypes==1.2.0
Jinja2==3.1.2
MarkupPy==1.14
MarkupSafe==2.1.3
odfpy==1.4.1
openpyxl==3.1.2
pycparser==2.21
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
sqlparse==0.4.4
tablib==3.5.0
uritemplate==4.1.1
urllib3==2.1.0
whitenoise==6.6.0
xlrd==2.0.1
xlwt==1.3.0
